LightOrb (+ Materials, quick addition to tempcharacter) added
I added the Light Orb Blueprint which contains some functions to increase, reduce and get the level of our light. Furthermore I added a function to set the current condition of the orb and its properties.
The Light Orb features a Point Light and a Collision Sphere, both of which get changed in the "Set Condition" function.
In addition to that Emissive Material for the Orb was added.

I tried to leave it as vague as possible so we/I can easily work on it later.