+ LightBar; + Gradual Light Reduction, + Respawnpoints, + LightDeath, - TurningCharacterWithMouse and a lot of bugfixing because of it.
Most of it will be explained in my comments in the code.
Adding the Light Orb gradual Reduction lead to me implementing the Light Orb related death, when it goes out. That led me to implement and sort of rework the Respawn UI, where I then needed to create a spawn point and thought why not make it to "Update the respawn point" during play. Currently they are small discs that need to be walked over. But because of that I had to revamp a lot, because it spawns in a new actor but did not "reactivate" the "BeginPlay"-Event in the ThirdPersonCharacter. And because of that I had to put the respawn into the ThirdPerson Gamemode to keep the variables and everything going.... Wooof. Now I am tired, but everything I checked/could check for should work atm.