Audio Placeholders placed in the level.
Audio cues(?) were placed inside of the level! You can check them out in the prototype level. They may get quite annoying, so you can turn their audio to 0 if you plan on playtesting the level, but in essence, both the inside and outside areas now have their own separate sounds.

The only thing I did NOT implement was audio occlusion, but I do understand how it works, so it should not be a problem once we do implement it.