Raise Mechanic added!
You can only raise the light if it has 2 Strength or more. Each use reduces the light slightly. you can hold it as long as you want. Same button as the control mechanic, which should activate automatically if you enter the range with a raised light and >= 3 Light Strength. Obviously the numbers and everything else can still be adjusted. In theory this deals damage in a small aoe if an enemy is at some point getting to close to the light. Possibly going to add that it only damages if it is raised while overheating or at higher strength.