EdgeGrab fixed
Edge grab works only on specific blocks which are allowed (you have to set this by clicking on a specific block and setting the collision by line trace thing to edge grab and in advanced to block instead of ignore). You can no longer move when you are on the ledge.