Adiitional changes to the dissolution of vantablack crystals.
The dissolution of vantablack works only partly now, and there are still some issues that need to be dealt with. For example, there can only ever be one instance of a dissolver in a scene, not multiple, this goes directly against the blow mechanic function, as it is always instantiated and we often times see many versions of it at the same time.

A solution to this would be to remove blow mechanic's instantiation completely -- this would serve to make it (in my eyes) more efficient as well as allow it to work much better with the dissolver actor.