sorted and adjusted stuff
in the sewer level everything should be sorted by sections. The first section is the start, first tube, water and everything around the water. second secion is the first pipe up and the pipe where u get out of after the ladder and the pipe opposite from it, as well as the crain part and platforms. third section is the vertical cylinder and everything inside that.