Testing out main menu and ui controller issues
Kind of got the ui to work as a test. when you click chapter select with mouse you can then take your controller and if you hold down A or bottom special button then you will be able to flip through the options and select, but its weird as to why i need to do that. also if you fuck around with it too much it disappears?