Startup video
Created a placeholder video which plays upon startup. This will only play in a standalone game or when the build is complete. The video is skippable and it will complete once the level is loaded up. This can be changed, I can change it not to be skippable and to finish completely. This is different for the loading screens